+++ MiniProgram/src/roomkit/TUIRoom/components/RoomContent/StreamRegion/index.vue
@@ -24,7 +24,9 @@
       <Avatar class="avatar-region" :img-src="stream.avatarUrl"></Avatar>
     </div>
     <div class="corner-user-info-container">
-      <svg-icon v-if="showMasterIcon" size="custom" class="master-icon" icon-name="user"></svg-icon>
+      <div v-if="showMasterIcon" class="master-icon">
+        <svg-icon style="display: flex" :icon="UserIcon"></svg-icon>
+      </div>
       <audio-icon
         v-if="!isScreenStream"
         class="audio-icon"
@@ -32,7 +34,7 @@
         :is-muted="!stream.hasAudioStream"
         size="small"
       ></audio-icon>
-      <svg-icon v-if="isScreenStream" icon-name="screen-share" class="screen-icon"></svg-icon>
+      <svg-icon style="display: flex" v-if="isScreenStream" :icon="ScreenOpenIcon" class="screen-icon"></svg-icon>
       <span class="user-name" :title="userInfo">{{ userInfo }}</span>
       <span v-if="isScreenStream"> {{ t('is sharing their screen') }} </span>
     </div>
@@ -42,11 +44,13 @@
 <script setup lang="ts">
 import { ref, watch, nextTick, computed, onMounted } from 'vue';
 import { StreamInfo, useRoomStore } from '../../../stores/room';
-import Avatar from '../../base/Avatar.vue';
+import Avatar from '../../common/Avatar.vue';
 import { useBasicStore } from '../../../stores/basic';
 import logger from '../../../utils/common/logger';
-import AudioIcon from '../../base/AudioIcon.vue';
-import SvgIcon from '../../common/SvgIcon.vue';
+import AudioIcon from '../../common/AudioIcon.vue';
+import SvgIcon from '../../common/base/SvgIcon.vue';
+import UserIcon from '../../../assets/icons/UserIcon.svg';
+import ScreenOpenIcon from '../../../assets/icons/ScreenOpenIcon.svg';
 import { useI18n } from '../../../locales';
 import {
   TUIVideoStreamType,
@@ -107,7 +111,6 @@ onMounted(() => {
           if (basicStore.userId === props.stream.userId) {
             if (props.stream.hasVideoStream) {
               await roomEngine.instance?.setLocalVideoView({
-                streamType: TUIVideoStreamType.kCameraStream,
                 view: `${playRegionDomId.value}`,
               });
             }
@@ -187,7 +190,6 @@ onMounted(() => {
           **/
           if (props.stream.hasVideoStream) {
             await roomEngine.instance?.setLocalVideoView({
-              streamType: TUIVideoStreamType.kCameraStream,
               view: `${playRegionDomId.value}`,
             });
           }
@@ -212,7 +214,6 @@ onMounted(() => {
 </script>
 
 <style lang="scss" scoped>
-@import '../../../assets/style/var.scss';
 .user-stream-container {
   position: absolute;
   top: 0;
@@ -226,7 +227,6 @@ onMounted(() => {
   -webkit-transform: rotate(0deg);
   &.border {
     border: 2px solid #37E858;
-    box-sizing: border-box;
   }
 
   .center-user-info-container {
@@ -281,7 +281,7 @@ onMounted(() => {
     height: 30px;
     display: flex;
     background: rgba(0,0,0,0.60);
-    color: $whiteColor;
+    color: #FFFFFF;
     align-items: center;
     align-content: center;
     font-size: 14px;
@@ -295,9 +295,12 @@ onMounted(() => {
     }
     .master-icon {
       margin-left: 0;
-      width: 25px;
-      height: 25px;
-      background-size: cover;
+      width: 32px;
+      height: 32px;
+      background-color: var(--active-color-1);
+      display: flex;
+      justify-content: center;
+      align-items: center;
     }
     .screen-icon {
       transform: scale(0.8);