+++ MiniProgram/src/roomkit/TUIRoom/components/RoomFooter/EndControl/index.vue
@@ -1,7 +1,10 @@
 <template>
   <div>
     <div class="end-control-container">
-      <div class="end-button" tabindex="1" @tap="stopMeeting">{{ t('EndH5') }}</div>
+      <div @tap="handleEndBtnClick" class="end-button" tabindex="1">
+        <svg-icon style="display: flex" :icon="EndRoomIcon" />
+        <span class="end-button-title">{{ t('EndH5') }}</span>
+      </div>
     </div>
     <div v-if="visible" class="end-main-content">
       <div :class="isShowLeaveRoomDialog ? 'end-dialog-leave':'end-dialog-dismiss'">
@@ -24,8 +27,8 @@
           </span>
           <span
             v-if="isShowLeaveRoomDialog"
+            @tap="handleEndLeaveClick"
             :class="roomStore.isMaster ?'end-button-leave':'end-button-leave-single'"
-            @tap="leaveRoom"
           >
             {{ t('Leave') }}
           </span>
@@ -64,11 +67,12 @@
                   <Avatar class="avatar-url" :img-src="user.avatarUrl"></Avatar>
                   <div class="user-name">{{ user.userName || user.userId }}</div>
                   <svg-icon
+                    style="display: flex"
                     v-if="selectedUser === user.userId"
+                    :icon="CorrectIcon"
                     class="correct"
-                    icon-name="correct"
-                    size="custom"
-                  ></svg-icon>
+                  >
+                  </svg-icon>
                 </div>
               </div>
               <div v-if="hasNoData" class="member-hasNoData">{{ t('No relevant user found.') }}</div>
@@ -87,13 +91,16 @@
 
 <script setup lang="ts">
 import { onUnmounted } from 'vue';
-import { ElMessageBox, ElMessage } from '../../../elementComp';
 import { TUIRoomEngine, TUIRole, TUIRoomEvents } from '@tencentcloud/tuiroom-engine-wx';
 import useEndControl from './useEndControlHooks';
 import logger from '../../../utils/common/logger';
-import popup from '../../common/PopUpH5.vue';
-import SvgIcon from '../../common/SvgIcon.vue';
-import Avatar from '../../base/Avatar.vue';
+import popup from '../../common/base/PopUpH5.vue';
+import SvgIcon from '../../common/base/SvgIcon.vue';
+import CorrectIcon from '../../../assets/icons/CorrectIcon.svg';
+import Avatar from '../../common/Avatar.vue';
+import EndRoomIcon from '../../../assets/icons/EndRoomIcon.svg';
+import TUIMessage from '../../common/base/Message/index';
+import TUIMessageBox from '../../common/base/MessageBox/index';
 
 const {
   t,
@@ -117,11 +124,38 @@ const {
   filteredList,
   selectedUser,
   showSideBar,
+  remoteAnchorList,
+  isMasterWithOneRemoteAnchor,
+  isMasterWithRemoteAnchors,
+  isMasterWithoutRemoteAnchors,
 } = useEndControl();
 
 
 const emit = defineEmits(['on-exit-room', 'on-destroy-room']);
-
+function handleEndBtnClick() {
+  if (isMasterWithoutRemoteAnchors.value) {
+    dismissRoom();
+  } else {
+    stopMeeting();
+  }
+}
+function handleEndLeaveClick() {
+  if (!roomStore.isMaster) {
+    leaveRoom();
+    return;
+  }
+  if (isMasterWithRemoteAnchors.value) {
+    selectedUser.value = remoteAnchorList.value[0].userId;
+    if (isMasterWithOneRemoteAnchor.value) {
+      transferAndLeave();
+      return;
+    }
+    currentDialogType.value = DialogType.TransferDialog;
+    toggleMangeMemberSidebar();
+    resetState();
+    return;
+  }
+}
 
 /**
  * Active room dismissal
@@ -147,11 +181,6 @@ async function dismissRoom() {
 **/
 async function leaveRoom() { // eslint-disable-line
   try {
-    if (roomStore.isMaster) {
-      toggleMangeMemberSidebar();
-      resetState();
-      return;
-    }
     await closeMediaBeforeLeave();
     const response = await roomEngine.instance?.exitRoom();
     logger.log(`${logPrefix}leaveRoom:`, response);
@@ -191,10 +220,11 @@ const onRoomDismissed = async (eventInfo: { roomId: string}) => {
   try {
     const { roomId } = eventInfo;
     logger.log(`${logPrefix}onRoomDismissed:`, roomId);
-    ElMessageBox.alert(t('The host closed the room.'), t('Note'), {
-      customClass: 'custom-element-class',
-      confirmButtonText: t('Confirm'),
-      appendTo: '#roomContainer',
+    TUIMessageBox({
+      title: t('Note'),
+      message: t('The host closed the room.'),
+      appendToRoomContainer: true,
+      confirmButtonText: t('Sure'),
       callback: async () => {
         resetState();
         emit('on-destroy-room', { code: 0, message: '' });
@@ -219,7 +249,7 @@ const onUserRoleChanged = async (eventInfo: {userId: string, userRole: TUIRole }
       newName = t('me');
     }
     const tipMessage = `${t('Moderator changed to ')}${newName}`;
-    ElMessage({
+    TUIMessage({
       type: 'success',
       message: tipMessage,
     });
@@ -249,11 +279,8 @@ onUnmounted(() => {
 
 </script>
 <style lang="scss" scoped>
-@import '../../../assets/style/var.scss';
 .end-control-container{
   .end-button {
-    border: 1px solid #FF2E2E;
-    border-radius: 4px;
     font-weight: 400;
     font-size: 12px;
     color: #FF2E2E;
@@ -261,10 +288,9 @@ onUnmounted(() => {
     cursor: pointer;
     text-align: center;
     line-height: 21px;
-    padding:  0 8px;
-    &:hover {
-      background-color: #FF2E2E;
-      color: $whiteColor;
+    display: flex;
+    .end-button-title {
+      margin-left: 3px;
     }
   }
 }
@@ -440,6 +466,7 @@ onUnmounted(() => {
     .member-basic-info {
       display: flex;
       position: relative;
+      width: 100%;
       flex-direction: row;
       align-items: center;
       padding: 24px 20px 0 24px;