+++ MiniProgram/src/roomkit/TUIRoom/components/RoomHeader/UserInfo/index.vue
@@ -36,8 +36,7 @@ interface Props {
 defineProps<Props>();
 defineEmits(['log-out']);
 </script>
-<style lang="scss">
-@import '../../../assets/style/var.scss';
+<style lang="scss" scoped>
 @import '../../../assets/style/element-custom.scss';
 
 .user-info-container {
@@ -58,7 +57,7 @@ defineEmits(['log-out']);
       white-space: nowrap;
       overflow: hidden;
       text-overflow: ellipsis;
-      color: var(--color-font);
+      color: var(--font-color);
     }
     .down-icon {
       margin-left: 4px;
@@ -67,7 +66,7 @@ defineEmits(['log-out']);
   .user-control-container {
     background: var(--user-control-container-color-bg);
     box-shadow: var(--user-control-container-shadow);
-    color: var(--color-font);
+    color: var(--font-color);
     padding: 10px 0;
     position: absolute;
     top: calc(100% + 14px);
@@ -93,7 +92,7 @@ defineEmits(['log-out']);
       background:  var(--log-out-mobile);
       &-main{
         position: absolute;
-        top: 85%;
+        bottom: 34px;
         width: 100%;
         display: flex;
         flex-direction: column;